requirements.txt
# Core Web Framework
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
websockets==14.1
starlette==0.48.0
sse-starlette==3.0.2

# AI & LLM Models
google-generativeai==0.8.3
groq==0.32.0
openai==2.3.0
anthropic==0.69.0

# Google APIs (Gmail, Sheets, etc)
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.35.0
google-api-core==2.26.0
googleapis-common-protos==1.70.0

# Web Automation & Scraping
playwright==1.49.1
requests==2.32.5
requests-oauthlib==2.0.0

# NLP & Text Processing
spacy==3.7.5
langcodes==3.5.0
srsly==2.5.1
thinc==8.2.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Data Processing & Serialization
pydantic==2.12.0
numpy==1.26.4
rich==14.2.0
markdown-it-py==4.0.0
mdurl==0.1.2

# System & Process Management
psutil>=5.9.0
distro==1.9.0
shellingham==1.5.4

# Screenshots & Display
pillow==11.3.0
mss==10.1.0
Pygments==2.19.2

# Hardware & Input
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyScreeze==1.0.1
pytweening==1.2.0
MouseInfo==0.1.3
PySocks==1.7.1

# Async & Concurrency
anyio==4.11.0
trio==0.31.0
trio-websocket==0.12.2
greenlet==3.1.1
sniffio==1.3.1
outcome==1.3.0.post0

# HTTP & Network
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
h11==0.16.0

# Authentication & Security
oauthlib==3.3.1
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Protocol Buffers & gRPC
protobuf==5.29.5
grpcio==1.75.1
grpcio-status==1.71.2

# Utilities & Configuration
python-dotenv==1.1.1
typing_extensions==4.15.0
typer==0.19.2
click==8.3.0
packaging==25.0

# Dependency Management
setuptools==80.9.0
attrs==25.4.0
annotated-types==0.7.0
pydantic_core==2.41.1

# Code Quality & Development
wrapt==1.17.3
jiter==0.11.0

# Misc Dependencies
six==1.17.0
python-xlib==0.33
tqdm==4.67.1
smart_open==7.3.1
sortedcontainers==2.4.0
wsproto==1.2.0
websocket-client==1.9.0
httpx-sse==0.4.3
pyee==12.0.0

# spaCy dependencies
catalogue==2.0.10
cymem==2.0.11
confection==0.1.5
murmurhash==1.0.13
preshed==3.0.10
wasabi==1.1.3
srsly==2.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
language_data==1.3.0
marisa-trie==1.3.1

# Advanced Automation System
SQLAlchemy==2.0.23
alembic==1.13.1
